Split all endpoints into dedicated router modules.
